backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase>=2.5.1
pydantic==2.6.0
python-dotenv==1.0.1
python-multipart==0.0.6
httpx>=0.27.0
lunar-python==1.3.11
replicate==0.25.0
anyio>=4.0.0
runwayml>=0.2.0
